+++ blocks/switch.py
@@ -1,6 +1,7 @@
 import numpy as np
 from blocks.base_block import BaseBlock
 from blocks.input_helpers import get_scalar
+from blocks.param_templates import zero_crossing_param
 
 
 class SwitchBlock(BaseBlock):
@@ -54,6 +55,7 @@ class SwitchBlock(BaseBlock):
                 "options": ["threshold", "index"],
                 "doc": "'threshold' or 'index'.",
             },
+            **zero_crossing_param(),
         }
 
     @property